+++ src/aspire/image/__init__.py
@@ -8,4 +8,3 @@ from .image import (
     _im_translate2,
     normalize_bg,
 )
-from .preprocess import crop_pad, downsample, fuzzy_mask
