+++ src/aspire/utils/misc.py
@@ -8,6 +8,7 @@ import sys
 from itertools import chain, combinations
 
 import numpy as np
+from scipy.special import erf
 
 from aspire.utils import grid_1d, grid_2d, grid_3d
 from aspire.utils.rotation import Rotation
@@ -255,6 +256,33 @@ def inverse_r(size, x0=0, y0=0, peak=1, dtype=np.float64):
     vals = np.sqrt(1 + (g["x"] - x0) ** 2 + (g["y"] - y0) ** 2)
 
     return (peak / vals).astype(dtype)
+
+
+def fuzzy_mask(L, r0, risetime, origin=None):
+    """
+    Create a centered 1D to 3D fuzzy mask of radius r0
+
+    Made with an error function with effective rise time.
+
+    :param L: The sizes of image in tuple structure
+    :param r0: The specified radius
+    :param risetime: The rise time for `erf` function
+    :param origin: The coordinates of origin
+    :return: The desired fuzzy mask
+    """
+
+    center = [sz // 2 + 1 for sz in L]
+    if origin is None:
+        origin = center
+
+    grids = [np.arange(1 - org, ell - org + 1) for ell, org in zip(L, origin)]
+    XYZ = np.meshgrid(*grids, indexing="ij")
+    XYZ_sq = [X**2 for X in XYZ]
+    R = np.sqrt(np.sum(XYZ_sq, axis=0))
+    k = 1.782 / risetime
+    m = 0.5 * (1 - erf(k * (R - r0)))
+
+    return m
 
 
 def all_pairs(n):
